# Reduce ObjectPool overhead