<!--  Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines: https://github.com/m3db/m3/blob/master/CONTRIBUTING.md and developer notes https://github.com/m3db/m3/blob/master/DEVELOPER.md
2. Please prefix the name of the pull request with the component you are updating in the format "[component] Change title" (for example "[dbnode] Support out of order writes") and also label this pull request according to what type of issue you are addressing. Furthermore, if this is a WIP or DRAFT PR, please create a draft PR instead: https://github.blog/2019-02-14-introducing-draft-pull-requests/
3. Ensure you have added or ran the appropriate tests for your PR: read more at https://github.com/m3db/m3/blob/master/DEVELOPER.md#testing-changes
4. Follow the instructions for writing a changelog note: read more at https://github.com/m3db/m3/blob/master/DEVELOPER.md#adding-a-changelog
-->

**What this PR does / why we need it**:
<!--
If you have an issue this change addresses, please add the following details:
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Add rudimentary benchmarks to measure impact of any changes to pooling
Reduce instrumentation and raw get/put overhead due to contention
```
name                                                 old time/op    new time/op    delta
ObjectPoolGetPut                                       68.3ns ± 3%    58.6ns ± 3%  -14.24%  (p=0.000 n=10+10)
ObjectPoolGetPut-2                                     68.0ns ± 1%    58.5ns ± 2%  -13.95%  (p=0.000 n=10+10)
ObjectPoolGetPut-4                                     68.5ns ± 2%    58.0ns ± 1%  -15.36%  (p=0.000 n=10+9)
ObjectPoolGetPut-6                                     68.6ns ± 3%    58.2ns ± 1%  -15.16%  (p=0.000 n=10+10)
ObjectPoolGetPut-8                                     68.3ns ± 2%    58.2ns ± 2%  -14.76%  (p=0.000 n=10+10)
ObjectPoolGetPut-12                                    68.5ns ± 3%    58.3ns ± 2%  -14.77%  (p=0.000 n=10+10)
ObjectPoolParallelGetPut                               73.4ns ± 2%    63.2ns ± 2%  -13.98%  (p=0.000 n=10+9)
ObjectPoolParallelGetPut-2                             86.0ns ± 1%    74.8ns ± 0%  -13.02%  (p=0.000 n=9+10)
ObjectPoolParallelGetPut-4                             97.7ns ± 1%    84.9ns ± 1%  -13.12%  (p=0.000 n=10+10)
ObjectPoolParallelGetPut-6                              112ns ± 3%      98ns ± 2%  -12.31%  (p=0.000 n=10+10)
ObjectPoolParallelGetPut-8                              143ns ± 2%     123ns ± 3%  -14.15%  (p=0.000 n=10+9)
ObjectPoolParallelGetPut-12                             176ns ± 5%     163ns ± 5%   -7.24%  (p=0.000 n=10+10)
ObjectPoolParallelGetMultiPutContended                 1.41µs ± 2%    1.28µs ± 6%   -8.96%  (p=0.000 n=9+10)
ObjectPoolParallelGetMultiPutContended-2               1.65µs ± 2%    1.47µs ± 1%  -10.82%  (p=0.000 n=10+9)
ObjectPoolParallelGetMultiPutContended-4               1.96µs ± 1%    1.72µs ± 1%  -11.97%  (p=0.000 n=10+9)
ObjectPoolParallelGetMultiPutContended-6               2.27µs ± 5%    1.95µs ± 1%  -14.32%  (p=0.000 n=10+9)
ObjectPoolParallelGetMultiPutContended-8               2.55µs ± 0%    2.28µs ± 1%  -10.64%  (p=0.000 n=8+10)
ObjectPoolParallelGetMultiPutContended-12              3.32µs ± 7%    3.11µs ± 6%   -6.35%  (p=0.002 n=10+10)
ObjectPoolParallelGetMultiPutContendedWithRefill       1.45µs ± 7%    1.30µs ± 6%  -10.29%  (p=0.000 n=10+10)
ObjectPoolParallelGetMultiPutContendedWithRefill-2     1.58µs ± 1%    1.41µs ± 1%  -10.98%  (p=0.000 n=9+10)
ObjectPoolParallelGetMultiPutContendedWithRefill-4     1.98µs ± 0%    1.63µs ± 3%  -17.57%  (p=0.000 n=9+10)
ObjectPoolParallelGetMultiPutContendedWithRefill-6     2.32µs ± 2%    1.93µs ± 7%  -16.74%  (p=0.000 n=9+10)
ObjectPoolParallelGetMultiPutContendedWithRefill-8     2.76µs ± 1%    2.32µs ± 4%  -15.96%  (p=0.000 n=10+10)
ObjectPoolParallelGetMultiPutContendedWithRefill-12    3.36µs ± 8%    3.10µs ± 7%   -7.75%  (p=0.002 n=10+10)
```
**Special notes for your reviewer**:

**Does this PR introduce a user-facing and/or backwards incompatible change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note

```

**Does this PR require updating code package or user-facing documentation?**:
<!--
If no, just write "NONE" in the documentation-note block below.
If yes, describe which documentation you updated.
Note: Any changes that significantly updates how a code package is architectured or functions requires code package documentation updates in the form of updates to the README.md file in that package.
-->
```documentation-note

```